Make endian single-header
